+++ src/test/java/com/mx/ntlink/client/NtLinkClientNominaTest.java
@@ -8,7 +8,6 @@ import java.nio.file.Paths;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import org.junit.Assert;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -720,17 +719,15 @@ public class NtLinkClientNominaTest {
     }
   }
 
-
-
   @Test
   public void timbra_nomina_error30() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM30.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM30.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -741,21 +738,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM30 - El nodo Nomina no se puede utilizar dentro del elemento ComplementoConcepto.",
-              e.getMessage());
+          "Request Error :NOM30 - El nodo Nomina no se puede utilizar dentro del elemento ComplementoConcepto.",
+          e.getMessage());
     }
   }
-
 
   @Test
   public void timbra_nomina_error31() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM31.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM31.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -766,22 +762,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM31 - El nodo Nomina no tiene valores en TotalPercepciones y/o TotalOtrosPagos.",
-              e.getMessage());
+          "Request Error :NOM31 - El nodo Nomina no tiene valores en TotalPercepciones y/o TotalOtrosPagos.",
+          e.getMessage());
     }
   }
-
-
 
   @Test
   public void timbra_nomina_error32() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM32.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM32.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -792,21 +786,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM32 - El valor del atributo Nomina.TipoNomina no cumple con un valor del catálogo c_TipoNomina.",
-              e.getMessage());
+          "Request Error :NOM32 - El valor del atributo Nomina.TipoNomina no cumple con un valor del catálogo c_TipoNomina.",
+          e.getMessage());
     }
   }
-
 
   @Test
   public void timbra_nomina_error33() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM33.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM33.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -817,8 +810,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM33 - El valor del atributo tipo de periodicidad es \"99\".",
-              e.getMessage());
+          "Request Error :NOM33 - El valor del atributo tipo de periodicidad es \"99\".",
+          e.getMessage());
     }
   }
 
@@ -826,11 +819,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error34() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM34.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM34.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -841,21 +834,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM34 - El valor del atributo tipo de periodicidad no tiene el valor \"99\".",
-              e.getMessage());
+          "Request Error :NOM34 - El valor del atributo tipo de periodicidad no tiene el valor \"99\".",
+          e.getMessage());
     }
   }
-
 
   @Test
   public void timbra_nomina_error35() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM35.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM35.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -866,21 +858,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM35 - El valor del atributo Nomina.FechaInicialPago no es menor o igual al valor del atributo FechaFinalPago.",
-              e.getMessage());
+          "Request Error :NOM35 - El valor del atributo Nomina.FechaInicialPago no es menor o igual al valor del atributo FechaFinalPago.",
+          e.getMessage());
     }
   }
-
 
   @Test
   public void timbra_nomina_error36() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM36.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM36.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -891,8 +882,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM36 - El atributo Nomina.TotalPercepciones, no debe existir.",
-              e.getMessage());
+          "Request Error :NOM36 - El atributo Nomina.TotalPercepciones, no debe existir.",
+          e.getMessage());
     }
   }
 
@@ -900,11 +891,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error37() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM37.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM37.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -915,8 +906,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM37 - El valor del atributo Nomina.TotalPercepciones no coincide con la suma de los atributos TotalSueldos más TotalSeparacionIndemnizacion más TotalJubilacionPensionRetiro del  nodo Percepciones.",
-              e.getMessage());
+          "Request Error :NOM37 - El valor del atributo Nomina.TotalPercepciones no coincide con la suma de los atributos TotalSueldos más TotalSeparacionIndemnizacion más TotalJubilacionPensionRetiro del  nodo Percepciones.",
+          e.getMessage());
     }
   }
 
@@ -924,11 +915,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error38() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM38.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM38.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -939,21 +930,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM38 - El atributo Nomina.TotalDeducciones, no debe existir.",
-              e.getMessage());
+          "Request Error :NOM38 - El atributo Nomina.TotalDeducciones, no debe existir.",
+          e.getMessage());
     }
   }
-
 
   @Test
   public void timbra_nomina_error39() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM39.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM39.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -964,8 +954,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM39 - El valor del atributo Nomina.TotalDeducciones no coincide con la suma de los atributos TotalOtrasDeducciones más TotalImpuestosRetenidos del elemento Deducciones.",
-              e.getMessage());
+          "Request Error :NOM39 - El valor del atributo Nomina.TotalDeducciones no coincide con la suma de los atributos TotalOtrasDeducciones más TotalImpuestosRetenidos del elemento Deducciones.",
+          e.getMessage());
     }
   }
 
@@ -973,11 +963,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error40() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM40.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM40.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -988,8 +978,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM40 - El valor del atributo Nomina.TotalOtrosPagos no está registrado o  no coincide con la suma de los atributos Importe de los nodos nomina12:OtrosPagos:OtroPago.",
-              e.getMessage());
+          "Request Error :NOM40 - El valor del atributo Nomina.TotalOtrosPagos no está registrado o  no coincide con la suma de los atributos Importe de los nodos nomina12:OtrosPagos:OtroPago.",
+          e.getMessage());
     }
   }
 
@@ -997,11 +987,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error41() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM41.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM41.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1012,8 +1002,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM41 - El atributo Nomina.Emisor.RfcPatronOrigen no está inscrito o bien, inscrito y sí cancelado en el SAT (l_RFC).",
-              e.getMessage());
+          "Request Error :NOM41 - El atributo Nomina.Emisor.RfcPatronOrigen no está inscrito o bien, inscrito y sí cancelado en el SAT (l_RFC).",
+          e.getMessage());
     }
   }
 
@@ -1021,11 +1011,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error42() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM42.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM42.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1036,8 +1026,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM42 - El atributo Nomina.Emisor.RegistroPatronal se debe registrar.",
-              e.getMessage());
+          "Request Error :NOM42 - El atributo Nomina.Emisor.RegistroPatronal se debe registrar.",
+          e.getMessage());
     }
   }
 
@@ -1045,11 +1035,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error43() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM43.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM43.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1060,21 +1050,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM43 - El atributo Nomina.Emisor.RegistroPatronal  no se debe registrar.",
-              e.getMessage());
+          "Request Error :NOM43 - El atributo Nomina.Emisor.RegistroPatronal  no se debe registrar.",
+          e.getMessage());
     }
   }
-
 
   @Test
   public void timbra_nomina_error44() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM44.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM44.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1085,21 +1074,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM44 - Los atributos NumSeguridadSocial, FechaInicioRelLaboral, Antigüedad,RiesgoPuesto,SalarioDiarioIntegrado deben existir.",
-              e.getMessage());
+          "Request Error :NOM44 - Los atributos NumSeguridadSocial, FechaInicioRelLaboral, Antigüedad,RiesgoPuesto,SalarioDiarioIntegrado deben existir.",
+          e.getMessage());
     }
   }
-
 
   @Test
   public void timbra_nomina_error45() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM45.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM45.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1110,8 +1098,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM45 - El elemento Nomina.Emisor.EntidadSNCF debe existir.",
-              e.getMessage());
+          "Request Error :NOM45 - El elemento Nomina.Emisor.EntidadSNCF debe existir.",
+          e.getMessage());
     }
   }
 
@@ -1119,11 +1107,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error46() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM46.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM46.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1134,8 +1122,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM46 - El elemento Nomina.Emisor.EntidadSNCF no debe existir.",
-              e.getMessage());
+          "Request Error :NOM46 - El elemento Nomina.Emisor.EntidadSNCF no debe existir.",
+          e.getMessage());
     }
   }
 
@@ -1143,11 +1131,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error47() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM47.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM47.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1158,8 +1146,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM47 - El valor del atributo Nomina.Emisor.EntidadSNCF.OrigenRecurso no cumple con un valor del catálogo c_OrigenRecurso.",
-              e.getMessage());
+          "Request Error :NOM47 - El valor del atributo Nomina.Emisor.EntidadSNCF.OrigenRecurso no cumple con un valor del catálogo c_OrigenRecurso.",
+          e.getMessage());
     }
   }
 
@@ -1167,11 +1155,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error48() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM48.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM48.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1182,8 +1170,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM48 - El atributo Nomina.Emisor.EntidadSNCF.MontoRecursoPropio debe existir.",
-              e.getMessage());
+          "Request Error :NOM48 - El atributo Nomina.Emisor.EntidadSNCF.MontoRecursoPropio debe existir.",
+          e.getMessage());
     }
   }
 
@@ -1191,11 +1179,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error49() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM49.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM49.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1206,21 +1194,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM49 - El atributo Nomina.Emisor.EntidadSNCF.MontoRecursoPropio no debe existir.",
-              e.getMessage());
+          "Request Error :NOM49 - El atributo Nomina.Emisor.EntidadSNCF.MontoRecursoPropio no debe existir.",
+          e.getMessage());
     }
   }
-
 
   @Test
   public void timbra_nomina_error50() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM50.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM50.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1231,19 +1218,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM50 - El valor del atributo Nomina.Emisor.EntidadSNCF.MontoRecursoPropio no es menor a la suma de los valores de los atributos TotalPercepciones y TotalOtrosPagos. .",
-              e.getMessage());
+          "Request Error :NOM50 - El valor del atributo Nomina.Emisor.EntidadSNCF.MontoRecursoPropio no es menor a la suma de los valores de los atributos TotalPercepciones y TotalOtrosPagos. .",
+          e.getMessage());
     }
   }
+
   @Test
   public void timbra_nomina_error51() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM51.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM51.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1254,21 +1242,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM51 - El valor del atributo Nomina.Receptor.TipoContrato no cumple con un valor del catálogo c_TipoContrato.",
-              e.getMessage());
+          "Request Error :NOM51 - El valor del atributo Nomina.Receptor.TipoContrato no cumple con un valor del catálogo c_TipoContrato.",
+          e.getMessage());
     }
   }
-
 
   @Test
   public void timbra_nomina_error52() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM52.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM52.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1279,8 +1266,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM52 - El valor del atributo Nomina.Receptor.TipoJornada no cumple con un valor del catálogo c_TipoJornada.",
-              e.getMessage());
+          "Request Error :NOM52 - El valor del atributo Nomina.Receptor.TipoJornada no cumple con un valor del catálogo c_TipoJornada.",
+          e.getMessage());
     }
   }
 
@@ -1288,11 +1275,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error53() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM53.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM53.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1303,21 +1290,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM53 - El valor del atributo Nomina.Receptor.FechaInicioRelLaboral no es menor o igual al atributo a FechaFinalPago.",
-              e.getMessage());
+          "Request Error :NOM53 - El valor del atributo Nomina.Receptor.FechaInicioRelLaboral no es menor o igual al atributo a FechaFinalPago.",
+          e.getMessage());
     }
   }
-
 
   @Test
   public void timbra_nomina_error54() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM54.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM54.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1328,8 +1314,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM54 - El valor numérico del atributo Nomina.Receptor.Antigüedad no es menor o igual al cociente de (la suma del número de días transcurridos entre la FechaInicioRelLaboral y la FechaFinalPago más uno) dividido entre siete.",
-              e.getMessage());
+          "Request Error :NOM54 - El valor numérico del atributo Nomina.Receptor.Antigüedad no es menor o igual al cociente de (la suma del número de días transcurridos entre la FechaInicioRelLaboral y la FechaFinalPago más uno) dividido entre siete.",
+          e.getMessage());
     }
   }
 
@@ -1337,11 +1323,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error55() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM55.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM55.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1352,8 +1338,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM55 - El valor del atributo Nomina.Receptor.Antigüedad. no cumple con el número de años, meses y días transcurridos entre la FechaInicioRelLaboral y la FechaFinalPago.",
-              e.getMessage());
+          "Request Error :NOM55 - El valor del atributo Nomina.Receptor.Antigüedad. no cumple con el número de años, meses y días transcurridos entre la FechaInicioRelLaboral y la FechaFinalPago.",
+          e.getMessage());
     }
   }
 
@@ -1361,11 +1347,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error56() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM56.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM56.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1376,8 +1362,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM56 - El valor del atributo Nomina.Receptor.TipoRegimen no cumple con un valor del catálogo c_TipoRegimen.",
-              e.getMessage());
+          "Request Error :NOM56 - El valor del atributo Nomina.Receptor.TipoRegimen no cumple con un valor del catálogo c_TipoRegimen.",
+          e.getMessage());
     }
   }
 
@@ -1385,11 +1371,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error57() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM57.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM57.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1400,21 +1386,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM57 - El valor del atributo Nomina.Receptor.TipoRegimen no es \"02\", \"03\" o \"04\".",
-              e.getMessage());
+          "Request Error :NOM57 - El valor del atributo Nomina.Receptor.TipoRegimen no es \"02\", \"03\" o \"04\".",
+          e.getMessage());
     }
   }
-
 
   @Test
   public void timbra_nomina_error58() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM58.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM58.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1425,8 +1410,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM58 - El valor del atributo Nomina.Receptor.TipoRegimen no está entre \"05\" a \"99\".",
-              e.getMessage());
+          "Request Error :NOM58 - El valor del atributo Nomina.Receptor.TipoRegimen no está entre \"05\" a \"99\".",
+          e.getMessage());
     }
   }
 
@@ -1434,11 +1419,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error59() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM59.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM59.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1449,8 +1434,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM59 - El valor del atributo Nomina.Receptor.RiesgoPuesto no cumple con un valor del catálogo c_RiesgoPuesto.",
-              e.getMessage());
+          "Request Error :NOM59 - El valor del atributo Nomina.Receptor.RiesgoPuesto no cumple con un valor del catálogo c_RiesgoPuesto.",
+          e.getMessage());
     }
   }
 
@@ -1458,11 +1443,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error60() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM60.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM60.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1473,8 +1458,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM60 - El valor del atributo Nomina.Receptor.PeriodicidadPago no cumple con un valor del catálogo c_PeriodicidadPago.",
-              e.getMessage());
+          "Request Error :NOM60 - El valor del atributo Nomina.Receptor.PeriodicidadPago no cumple con un valor del catálogo c_PeriodicidadPago.",
+          e.getMessage());
     }
   }
 
@@ -1482,11 +1467,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error61() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM61.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM61.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1497,21 +1482,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM61 - El valor del atributo Nomina.Receptor.Banco no cumple con un valor del catálogo c_Banco.",
-              e.getMessage());
+          "Request Error :NOM61 - El valor del atributo Nomina.Receptor.Banco no cumple con un valor del catálogo c_Banco.",
+          e.getMessage());
     }
   }
-
 
   @Test
   public void timbra_nomina_error62() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM62.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM62.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1522,21 +1506,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM62 - El atributo Nomina.Receptor.CuentaBancaria no cumple con la longitud de 10, 11, 16 o 18 posiciones.",
-              e.getMessage());
+          "Request Error :NOM62 - El atributo Nomina.Receptor.CuentaBancaria no cumple con la longitud de 10, 11, 16 o 18 posiciones.",
+          e.getMessage());
     }
   }
-
 
   @Test
   public void timbra_nomina_error63() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM63.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM63.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1547,8 +1530,7 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM63 - El atributo Banco no debe existir.",
-              e.getMessage());
+          "Request Error :NOM63 - El atributo Banco no debe existir.", e.getMessage());
     }
   }
 
@@ -1556,11 +1538,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error64() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM64.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM64.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1571,8 +1553,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM64 - El dígito de control del atributo CLABE no es correcto.",
-              e.getMessage());
+          "Request Error :NOM64 - El dígito de control del atributo CLABE no es correcto.",
+          e.getMessage());
     }
   }
 
@@ -1580,11 +1562,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error65() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM65.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM65.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1594,9 +1576,7 @@ public class NtLinkClientNominaTest {
     try {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
-      Assert.assertEquals(
-              "Request Error :NOM65 - El atributo Banco debe existir.",
-              e.getMessage());
+      Assert.assertEquals("Request Error :NOM65 - El atributo Banco debe existir.", e.getMessage());
     }
   }
 
@@ -1604,11 +1584,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error66() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM66.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM66.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1619,8 +1599,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM66 - El valor del atributo Nomina.Receptor.ClaveEntFed no cumple con un valor del catálogo c_Estado.",
-              e.getMessage());
+          "Request Error :NOM66 - El valor del atributo Nomina.Receptor.ClaveEntFed no cumple con un valor del catálogo c_Estado.",
+          e.getMessage());
     }
   }
 
@@ -1628,11 +1608,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error67() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM67.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM67.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1643,8 +1623,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM67 - El valor del atributo Nomina.Receptor.SubContratacion.RfcLabora no está en la lista de RFC (l_RFC).",
-              e.getMessage());
+          "Request Error :NOM67 - El valor del atributo Nomina.Receptor.SubContratacion.RfcLabora no está en la lista de RFC (l_RFC).",
+          e.getMessage());
     }
   }
 
@@ -1652,11 +1632,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error68() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM68.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM68.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1667,21 +1647,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM68 - La suma de los valores registrados en el atributo Nomina.Receptor.SubContratacion.PorcentajeTiempo no es igual a 100.",
-              e.getMessage());
+          "Request Error :NOM68 - La suma de los valores registrados en el atributo Nomina.Receptor.SubContratacion.PorcentajeTiempo no es igual a 100.",
+          e.getMessage());
     }
   }
-
 
   @Test
   public void timbra_nomina_error69() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM69.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM69.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1692,21 +1671,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM69 - La suma de los valores de los atributos TotalSueldos más TotalSeparacionIndemnizacion más TotalJubilacionPensionRetiro no es igual a la suma de los valores de los atributos TotalGravado más TotalExento.",
-              e.getMessage());
+          "Request Error :NOM69 - La suma de los valores de los atributos TotalSueldos más TotalSeparacionIndemnizacion más TotalJubilacionPensionRetiro no es igual a la suma de los valores de los atributos TotalGravado más TotalExento.",
+          e.getMessage());
     }
   }
-
 
   @Test
   public void timbra_nomina_error70() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM70.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM70.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1717,21 +1695,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM70 - El valor del atributo Nomina.Percepciones.TotalSueldos, no es igual a la suma de los atributos ImporteGravado e ImporteExento donde la clave expresada en el atributo TipoPercepcion es distinta de \"022\", \"023\",\"025\", \"039\" y \"044\".",
-              e.getMessage());
+          "Request Error :NOM70 - El valor del atributo Nomina.Percepciones.TotalSueldos, no es igual a la suma de los atributos ImporteGravado e ImporteExento donde la clave expresada en el atributo TipoPercepcion es distinta de \"022\", \"023\",\"025\", \"039\" y \"044\".",
+          e.getMessage());
     }
   }
-
 
   @Test
   public void timbra_nomina_error71() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM71.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM71.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1742,23 +1719,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM71 - El valor del atributo Nomina.Percepciones.TotalSeparacionIndemnizacion, no es igual a la suma de los atributos ImporteGravado e ImporteExento donde la clave en el atributo TipoPercepcion es igual a \"022\", \"023\" o \"025\".",
-              e.getMessage());
+          "Request Error :NOM71 - El valor del atributo Nomina.Percepciones.TotalSeparacionIndemnizacion, no es igual a la suma de los atributos ImporteGravado e ImporteExento donde la clave en el atributo TipoPercepcion es igual a \"022\", \"023\" o \"025\".",
+          e.getMessage());
     }
   }
-
-
-
 
   @Test
   public void timbra_nomina_error72() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM72.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM72.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1769,8 +1743,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM72 - El valor del atributo Nomina.Percepciones.TotalJubilacionPensionRetiro, no es igual a la suma de los atributos ImporteGravado e importeExento donde la clave expresada en el atributo TipoPercepcion es igual a \"039\" o \"044\".",
-              e.getMessage());
+          "Request Error :NOM72 - El valor del atributo Nomina.Percepciones.TotalJubilacionPensionRetiro, no es igual a la suma de los atributos ImporteGravado e importeExento donde la clave expresada en el atributo TipoPercepcion es igual a \"039\" o \"044\".",
+          e.getMessage());
     }
   }
 
@@ -1778,11 +1752,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error73() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM73.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM73.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1793,21 +1767,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM73 - El valor del atributo Nomina.Percepciones.TotalGravado, no es igual a la suma de los atributos ImporteGravado de los nodos Percepcion.",
-              e.getMessage());
+          "Request Error :NOM73 - El valor del atributo Nomina.Percepciones.TotalGravado, no es igual a la suma de los atributos ImporteGravado de los nodos Percepcion.",
+          e.getMessage());
     }
   }
-
 
   @Test
   public void timbra_nomina_error74() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM74.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM74.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1818,8 +1791,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM74 - El valor del atributo Nomina.Percepciones.TotalExento, no es igual a la suma de los atributos ImporteExento de los nodos Percepcion.",
-              e.getMessage());
+          "Request Error :NOM74 - El valor del atributo Nomina.Percepciones.TotalExento, no es igual a la suma de los atributos ImporteExento de los nodos Percepcion.",
+          e.getMessage());
     }
   }
 
@@ -1827,11 +1800,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error75() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM75.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM75.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1842,8 +1815,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM75 - La suma de los importes de los atributos ImporteGravado e ImporteExento no es mayor que cero.",
-              e.getMessage());
+          "Request Error :NOM75 - La suma de los importes de los atributos ImporteGravado e ImporteExento no es mayor que cero.",
+          e.getMessage());
     }
   }
 
@@ -1851,11 +1824,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error76() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM76.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM76.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1866,8 +1839,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM76 - El valor del atributo Nomina.Percepciones.Percepcion.TipoPercepcion no cumple con un valor del catálogo c_TipoPercepcion.",
-              e.getMessage());
+          "Request Error :NOM76 - El valor del atributo Nomina.Percepciones.Percepcion.TipoPercepcion no cumple con un valor del catálogo c_TipoPercepcion.",
+          e.getMessage());
     }
   }
 
@@ -1875,11 +1848,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error77() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM77.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM77.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1890,22 +1863,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM77 - El atributo TotalSueldos, debe existir, ya que la clave expresada en Nomina.Percepciones.Percepcion.TipoPercepcion es distinta de \"022\", \"023\", \"025\", \"039\" y \"044\".",
-              e.getMessage());
+          "Request Error :NOM77 - El atributo TotalSueldos, debe existir, ya que la clave expresada en Nomina.Percepciones.Percepcion.TipoPercepcion es distinta de \"022\", \"023\", \"025\", \"039\" y \"044\".",
+          e.getMessage());
     }
   }
-
-
 
   @Test
   public void timbra_nomina_error78() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM78.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM78.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1916,21 +1887,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM78 - El atributo TotalSeparacionIndemnizacion y el elemento SeparacionIndemnizacion, deben existir, ya que la clave expresada en Nomina.Percepciones.Percepcion.TipoPercepcion es \"022\" o \"023\" o \"025\".",
-              e.getMessage());
+          "Request Error :NOM78 - El atributo TotalSeparacionIndemnizacion y el elemento SeparacionIndemnizacion, deben existir, ya que la clave expresada en Nomina.Percepciones.Percepcion.TipoPercepcion es \"022\" o \"023\" o \"025\".",
+          e.getMessage());
     }
   }
-
 
   @Test
   public void timbra_nomina_error79() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM79.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM79.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1941,21 +1911,20 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM79 - El atributo TotalJubilacionPensionRetiro y el elemento JubilacionPensionRetiro deben existir,  ya que la clave expresada en el atributo Nomina.Percepciones.Percepcion.TipoPercepcion es \"039\" o \"044\".",
-              e.getMessage());
+          "Request Error :NOM79 - El atributo TotalJubilacionPensionRetiro y el elemento JubilacionPensionRetiro deben existir,  ya que la clave expresada en el atributo Nomina.Percepciones.Percepcion.TipoPercepcion es \"039\" o \"044\".",
+          e.getMessage());
     }
   }
-
 
   @Test
   public void timbra_nomina_error80() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM80.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM80.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1966,8 +1935,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM80 - El atributo TotalUnaExhibicion debe existir y no deben existir TotalParcialidad, MontoDiario, ya que la clave expresada en el atributo Nomina.Percepciones.Percepcion.TipoPercepcion es \"039\".",
-              e.getMessage());
+          "Request Error :NOM80 - El atributo TotalUnaExhibicion debe existir y no deben existir TotalParcialidad, MontoDiario, ya que la clave expresada en el atributo Nomina.Percepciones.Percepcion.TipoPercepcion es \"039\".",
+          e.getMessage());
     }
   }
 
@@ -1975,11 +1944,11 @@ public class NtLinkClientNominaTest {
   public void timbra_nomina_error81() throws IOException, SoapClientException {
 
     String comprobante =
-            new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM81.xml")));
+        new String(Files.readAllBytes(Paths.get("./src/test/resources/nomina-error/NOM81.xml")));
 
     comprobante =
-            comprobante.replace(
-                    DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
+        comprobante.replace(
+            DATE_REPLACEMENT, formatter.format(LocalDateTime.now().minusMinutes(10)));
 
     TimbraCfdiSinSello cfdiSinSello = new TimbraCfdiSinSello();
     cfdiSinSello.setPassword(TEST_PASS);
@@ -1990,10 +1959,8 @@ public class NtLinkClientNominaTest {
       client.timbrarSinSello(cfdiSinSello);
     } catch (SoapClientException e) {
       Assert.assertEquals(
-              "Request Error :NOM81 - El atributo TotalUnaExhibicion no debe existir y deben existir TotalParcialidad, MontoDiario, ya que la clave expresada en el atributo Nomina.Percepciones.Percepcion.TipoPercepcion es \"044\".",
-              e.getMessage());
+          "Request Error :NOM81 - El atributo TotalUnaExhibicion no debe existir y deben existir TotalParcialidad, MontoDiario, ya que la clave expresada en el atributo Nomina.Percepciones.Percepcion.TipoPercepcion es \"044\".",
+          e.getMessage());
     }
   }
-
-
 }
